Alignment is completed
Alignment is completed. It works for lambda file, with jaccard sim = 0.62. It needs a bit of tooning to upgrate the score, but so far so good.

It completely ignores unaligned fragments as they are very hard to deduct. Some of the aligned fragment output great errors in mutatons those are ignored as well. But for some cases some aligned (3 or 4) fragments have no shared positions.